requirements.txt
requests~=2.25.1
pandas~=1.2.4
Flask~=1.1.2
psycopg2-binary==2.8.6
gunicorn==20.1.0